Uplift of #13124 (squashed) to release
